backend/requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pandas==2.2.3
# db
neo4j==5.17.0

google-genai>=1.0.0
scrapfly-sdk>=0.8.8
beautifulsoup4>=4.12.3
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
httpx>=0.26.0

# auth
authlib>=1.3.0
itsdangerous>=2.2.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9